Update option tests
