+++ spec/GHC/Plugin/OllamaHoles/Options/Spec.hs
@@ -2,143 +2,330 @@
 
 module GHC.Plugin.OllamaHoles.Options.Spec (tests) where
 
+import Data.Aeson (Value(..))
+import Data.Aeson.KeyMap qualified as KM
 import Test.Tasty
 import Test.Tasty.HUnit
 
 import GHC.Plugin.OllamaHoles
-    ( mkTemplateSpec
-    )
+  ( mkTemplateSpec
+  )
 import GHC.Plugin.OllamaHoles.Options
-    ( Flags(..)
-    , defaultFlags
-    , parseFlags
-    )
+  ( Flags(..)
+  , defaultFlags
+  , parseCommandLineOptions
+  , OptError(..)
+  , Token(..)
+  )
 import GHC.Plugin.OllamaHoles.Template
-    ( TemplateSpec(..)
-    , TemplateSource(..)
-    , TemplateError(..)
-    , unsafeCreateRawTemplateName
-    )
+  ( TemplateSpec(..)
+  , TemplateSource(..)
+  , TemplateError(..)
+  , unsafeCreateRawTemplateName
+  )
 
 tests :: TestTree
-tests = testGroup "template-related option parsing"
-    [ parseFlagsTests
+tests =
+  testGroup "Options"
+    [ parserDefaultsTests
+    , parserSimpleTests
+    , parserPrecedenceTests
+    , parserFailureTests
+    , templateParserTests
     , mkTemplateSpecTests
     ]
 
-parseFlagsTests :: TestTree
-parseFlagsTests = testGroup "parseFlags"
+expectParseOk :: [String] -> IO (Flags, [Token])
+expectParseOk opts =
+  case parseCommandLineOptions defaultFlags opts of
+    Left err ->
+      assertFailure ("unexpected parse error: " <> show err) >> fail "unreachable"
+    Right ok ->
+      pure ok
+
+parserDefaultsTests :: TestTree
+parserDefaultsTests =
+  testGroup "parseCommandLineOptions defaults"
+    [ testCase "empty options yields defaultFlags and no unknowns" $ do
+        (flags, unknowns) <- expectParseOk []
+        flags @?= defaultFlags
+        unknowns @?= []
+    ]
+
+parserSimpleTests :: TestTree
+parserSimpleTests =
+  testGroup "parseCommandLineOptions simple options"
+    [ testCase "model= sets model_name" $ do
+        (flags, unknowns) <- expectParseOk ["model=phi4"]
+        model_name flags @?= "phi4"
+        unknowns @?= []
+
+    , testCase "backend= sets backend_name" $ do
+        (flags, unknowns) <- expectParseOk ["backend=openai"]
+        backend_name flags @?= "openai"
+        unknowns @?= []
+
+    , testCase "openai_base_url= sets openai_base_url" $ do
+        (flags, unknowns) <- expectParseOk ["openai_base_url=https://example.com/v1"]
+        openai_base_url flags @?= "https://example.com/v1"
+        unknowns @?= []
+
+    , testCase "openai_key_name= sets openai_key_name" $ do
+        (flags, unknowns) <- expectParseOk ["openai_key_name=MY_API_KEY"]
+        openai_key_name flags @?= "MY_API_KEY"
+        unknowns @?= []
+
+    , testCase "debug enables debug" $ do
+        (flags, unknowns) <- expectParseOk ["debug"]
+        debug flags @?= True
+        unknowns @?= []
+
+    , testCase "include-docs enables include_docs" $ do
+        (flags, unknowns) <- expectParseOk ["include-docs"]
+        include_docs flags @?= True
+        unknowns @?= []
+
+    , testCase "n= sets num_expr" $ do
+        (flags, unknowns) <- expectParseOk ["n=17"]
+        num_expr flags @?= 17
+        unknowns @?= []
+
+    , testCase "model-options= parses valid JSON object" $ do
+        (flags, unknowns) <- expectParseOk ["model-options={\"temperature\":1.0,\"num_ctx\":32000}"]
+        unknowns @?= []
+        case model_options flags of
+          Nothing ->
+            assertFailure "expected model_options to be set"
+          Just (Object obj) -> do
+            assertBool "expected temperature key" ("temperature" `KM.member` obj)
+            assertBool "expected num_ctx key" ("num_ctx" `KM.member` obj)
+          Just other ->
+            assertFailure ("expected JSON object, got: " <> show other)
+
+    , testCase "unknown options are reported and do not change flags" $ do
+        (flags, unknowns) <- expectParseOk ["bogus-option", "another=thing"]
+        flags @?= defaultFlags
+        unknowns @?= [BooleanToken "bogus-option", ValueToken "another" "thing"]
+
+    , testCase "unknown options are reported but do not block recognized options" $ do
+        (flags, unknowns) <- expectParseOk ["bogus-option", "model=qwen3", "debug"]
+        model_name flags @?= "qwen3"
+        debug flags @?= True
+        unknowns @?= [BooleanToken "bogus-option"]
+    ]
+
+parserPrecedenceTests :: TestTree
+parserPrecedenceTests =
+  testGroup "parseCommandLineOptions precedence"
+    [ testCase "leftmost model= wins" $ do
+        (flags, unknowns) <- expectParseOk ["model=first", "model=second"]
+        model_name flags @?= "first"
+        unknowns @?= []
+
+    , testCase "leftmost backend= wins" $ do
+        (flags, unknowns) <- expectParseOk ["backend=ollama", "backend=openai"]
+        backend_name flags @?= "ollama"
+        unknowns @?= []
+
+    , testCase "leftmost openai_base_url= wins" $ do
+        (flags, unknowns) <- expectParseOk
+          [ "openai_base_url=https://first.example"
+          , "openai_base_url=https://second.example"
+          ]
+        openai_base_url flags @?= "https://first.example"
+        unknowns @?= []
+
+    , testCase "leftmost openai_key_name= wins" $ do
+        (flags, unknowns) <- expectParseOk
+          [ "openai_key_name=FIRST_KEY"
+          , "openai_key_name=SECOND_KEY"
+          ]
+        openai_key_name flags @?= "FIRST_KEY"
+        unknowns @?= []
+
+    , testCase "leftmost n= wins" $ do
+        (flags, unknowns) <- expectParseOk ["n=3", "n=9"]
+        num_expr flags @?= 3
+        unknowns @?= []
+
+    , testCase "boolean flags are sticky" $ do
+        (flags, unknowns) <- expectParseOk ["debug", "debug", "include-docs", "include-docs"]
+        debug flags @?= True
+        include_docs flags @?= True
+        unknowns @?= []
+
+    , testCase "leftmost model-options= wins" $ do
+        (flags, unknowns) <- expectParseOk
+          [ "model-options={\"temperature\":0.1}"
+          , "model-options={\"temperature\":0.9}"
+          ]
+        unknowns @?= []
+        case model_options flags of
+          Nothing ->
+            assertFailure "expected model_options to be set"
+          Just (Object obj) ->
+            assertBool "expected temperature key from leftmost option"
+              ("temperature" `KM.member` obj)
+          Just other ->
+            assertFailure ("expected JSON object, got: " <> show other)
+
+    , testCase "leftmost wins even when interleaved with other options" $ do
+        (flags, unknowns) <- expectParseOk
+          [ "model=first"
+          , "debug"
+          , "backend=ollama"
+          , "model=second"
+          , "include-docs"
+          , "backend=openai"
+          ]
+        model_name flags @?= "first"
+        backend_name flags @?= "ollama"
+        debug flags @?= True
+        include_docs flags @?= True
+        unknowns @?= []
+    ]
+
+parserFailureTests :: TestTree
+parserFailureTests =
+  testGroup "parseCommandLineOptions failures"
+    [ testCase "invalid n returns structured error" $ do
+        parseCommandLineOptions defaultFlags ["n=not-an-int"]
+          @?= Left (InvalidInt "n" "not-an-int")
+
+    , testCase "invalid model-options returns structured error" $ do
+        case parseCommandLineOptions defaultFlags ["model-options={not json}"] of
+          Left (InvalidJson "model-options" "{not json}" _) ->
+            pure ()
+          other ->
+            assertFailure ("expected InvalidJson, got: " <> show other)
+    ]
+
+templateParserTests :: TestTree
+templateParserTests =
+  testGroup "parseCommandLineOptions template options"
     [ testCase "defaults contain no template path or name" $ do
-        let flags = parseFlags []
+        (flags, unknowns) <- expectParseOk []
         template_path flags @?= Nothing
         template_name flags @?= Nothing
         template_search_dir flags @?= "."
+        unknowns @?= []
 
     , testCase "template= sets path and clears name" $ do
-        let flags = parseFlags ["template=/tmp/prompt.txt"]
+        (flags, unknowns) <- expectParseOk ["template=/tmp/prompt.txt"]
         template_path flags @?= Just "/tmp/prompt.txt"
         template_name flags @?= Nothing
+        unknowns @?= []
 
     , testCase "template-name= sets name and clears path" $ do
-        let flags = parseFlags ["template-name=qwen"]
+        (flags, unknowns) <- expectParseOk ["template-name=qwen"]
         template_path flags @?= Nothing
         template_name flags @?= Just "qwen"
+        unknowns @?= []
 
     , testCase "template-dir= sets search dir" $ do
-        let flags = parseFlags ["template-dir=/tmp/templates"]
+        (flags, unknowns) <- expectParseOk ["template-dir=/tmp/templates"]
         template_search_dir flags @?= "/tmp/templates"
+        unknowns @?= []
 
     , testCase "leftmost template selector wins: path then name keeps path" $ do
-        let flags = parseFlags
-              [ "template=/tmp/a.txt"
-              , "template-name=qwen"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template=/tmp/a.txt"
+          , "template-name=qwen"
+          ]
         template_path flags @?= Just "/tmp/a.txt"
         template_name flags @?= Nothing
+        unknowns @?= []
 
     , testCase "leftmost template selector wins: name then path keeps name" $ do
-        let flags = parseFlags
-              [ "template-name=qwen"
-              , "template=/tmp/a.txt"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template-name=qwen"
+          , "template=/tmp/a.txt"
+          ]
         template_path flags @?= Nothing
         template_name flags @?= Just "qwen"
+        unknowns @?= []
 
     , testCase "leftmost template-dir wins" $ do
-        let flags = parseFlags
-              [ "template-dir=/tmp/one"
-              , "template-dir=/tmp/two"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/one"
+          , "template-dir=/tmp/two"
+          ]
         template_search_dir flags @?= "/tmp/one"
+        unknowns @?= []
 
     , testCase "template-dir combines with template-name" $ do
-        let flags = parseFlags
-              [ "template-dir=/tmp/templates"
-              , "template-name=qwen"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates"
+          , "template-name=qwen"
+          ]
         template_search_dir flags @?= "/tmp/templates"
         template_name flags @?= Just "qwen"
         template_path flags @?= Nothing
+        unknowns @?= []
 
     , testCase "template-dir combines with template path" $ do
-        let flags = parseFlags
-              [ "template-dir=/tmp/templates"
-              , "template=/tmp/prompt.txt"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates"
+          , "template=/tmp/prompt.txt"
+          ]
         template_search_dir flags @?= "/tmp/templates"
         template_path flags @?= Just "/tmp/prompt.txt"
         template_name flags @?= Nothing
+        unknowns @?= []
 
     , testCase "leftmost template selector still wins when interleaved with dir flags" $ do
-        let flags = parseFlags
-              [ "template-name=alpha"
-              , "template-dir=/tmp/one"
-              , "template=/tmp/prompt.txt"
-              , "template-dir=/tmp/two"
-              ]
+        (flags, unknowns) <- expectParseOk
+          [ "template-name=alpha"
+          , "template-dir=/tmp/one"
+          , "template=/tmp/prompt.txt"
+          , "template-dir=/tmp/two"
+          ]
         template_name flags @?= Just "alpha"
         template_path flags @?= Nothing
         template_search_dir flags @?= "/tmp/one"
+        unknowns @?= []
     ]
 
 mkTemplateSpecTests :: TestTree
-mkTemplateSpecTests = testGroup "mkTemplateSpec"
+mkTemplateSpecTests =
+  testGroup "mkTemplateSpec"
     [ testCase "default flags choose DefaultTemplate" $ do
         mkTemplateSpec defaultFlags
-            @?= Right (TemplateSpec
+          @?= Right (TemplateSpec
                 { tsSearchDir = "."
                 , tsSource = DefaultTemplate
                 })
 
     , testCase "path chooses TemplateFile" $ do
-        let flags = defaultFlags
-              { template_path = Just "/tmp/prompt.txt"
-              , template_name = Nothing
-              , template_search_dir = "/tmp/templates"
-              }
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates"
+          , "template=/tmp/prompt.txt"
+          ]
+        unknowns @?= []
         mkTemplateSpec flags
-            @?= Right (TemplateSpec
+          @?= Right (TemplateSpec
                 { tsSearchDir = "/tmp/templates"
                 , tsSource = TemplateFile "/tmp/prompt.txt"
                 })
 
     , testCase "name chooses NamedTemplate" $ do
-        let flags = defaultFlags
-              { template_path = Nothing
-              , template_name = Just "qwen"
-              , template_search_dir = "/tmp/templates"
-              }
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates"
+          , "template-name=qwen"
+          ]
+        unknowns @?= []
         mkTemplateSpec flags
-            @?= Right (TemplateSpec
+          @?= Right (TemplateSpec
                 { tsSearchDir = "/tmp/templates"
-                , tsSource = NamedTemplate $ unsafeCreateRawTemplateName "qwen"
+                , tsSource = NamedTemplate (unsafeCreateRawTemplateName "qwen")
                 })
 
     , testCase "search dir is preserved with default template" $ do
-        let flags = defaultFlags
-              { template_search_dir = "/tmp/templates" }
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates" ]
+        unknowns @?= []
         mkTemplateSpec flags
-            @?= Right (TemplateSpec
+          @?= Right (TemplateSpec
                 { tsSearchDir = "/tmp/templates"
                 , tsSource = DefaultTemplate
                 })
@@ -150,17 +337,17 @@ mkTemplateSpecTests = testGroup "mkTemplateSpec"
               , template_search_dir = "/tmp/templates"
               }
         mkTemplateSpec flags
-            @?= Right (TemplateSpec
+          @?= Right (TemplateSpec
                 { tsSearchDir = "/tmp/templates"
                 , tsSource = TemplateFile "/tmp/prompt.txt"
                 })
 
     , testCase "invalid name is rejected" $ do
-        let flags = defaultFlags
-              { template_path = Nothing
-              , template_name = Just "../secrets"
-              , template_search_dir = "/tmp/templates"
-              }
+        (flags, unknowns) <- expectParseOk
+          [ "template-dir=/tmp/templates"
+          , "template-name=../secrets"
+          ]
+        unknowns @?= []
         mkTemplateSpec flags
-            @?= Left (InvalidTemplateName "../secrets")
+          @?= Left (InvalidTemplateName "../secrets")
     ]
