+++ spec/Main.hs
@@ -9,7 +9,6 @@ import qualified GHC.Plugin.OllamaHoles.Template.Parse.Spec as ParseSpec
 import qualified GHC.Plugin.OllamaHoles.Template.Load.Spec as LoadSpec
 import qualified GHC.Plugin.OllamaHoles.Template.Expand.Spec as ExpandSpec
 import qualified GHC.Plugin.OllamaHoles.Options.Spec as OptionsSpec
-import qualified GHC.Plugin.OllamaHoles.Options.Parse.Spec as OptParseSpec
 
 main :: IO ()
 main = defaultMain $
@@ -22,5 +21,4 @@ main = defaultMain $
         , LoadSpec.tests
         , ExpandSpec.tests
         , OptionsSpec.tests
-        , OptParseSpec.tests
         ]
